+++ app/Http/Controllers/PaymentController.php
@@ -276,11 +276,15 @@ class PaymentController extends Controller
 
     public function destroy(Payment $payment)
     {
-        $text = ucwords(auth()->user()->name) . " deleted Payment : " . $payment->payment_number . ", datetime :   " . now();
+        if ($payment->can_delete()) {
+            $text = ucwords(auth()->user()->name) . " deleted Payment : " . $payment->payment_number . ", datetime :   " . now();
 
-        Log::create(['text' => $text]);
-        $payment->delete();
+            Log::create(['text' => $text]);
+            $payment->delete();
 
-        return redirect()->back()->with('error', 'Payment deleted successfully!');
+            return redirect()->back()->with('error', 'Payment deleted successfully!');
+        } else {
+            return redirect()->back()->with('error', 'Unable to delete...');
+        }
     }
 }
